requirements.txt
# requirements.txt
python-telegram-bot==13.14
python-dotenv==1.0.0
pymongo==4.3.3
pandas==1.5.3
matplotlib==3.7.1
pytz==2023.3 